Glass And Water made and imported
the base water shader has been made and in the scene. The glass texture is also there and put into a material which is ready for use. Some post processing/lighting has also been setup within the scene. Basic terrain has been made to test the water shader but is REMOVABLE WHENEVER WE WANT. Please don't hesitate to get rid of the terrain from this branch.